fix(server): remove optional catch-all route
